// src/app/core/model/ApproveMatrix/ViewApproveMatrix.ts
export interface IViewApproveMatrixEntity {
Id: number;
  MenuName: string;
  SubMenu: string;
  IsAllowedApproval: boolean;
  GroupName: string;
  Enroll: number;
  Sequence: string;
  Unit: string;
  IsActive: boolean;
  CreatedBy: number;
  CreatedDate: string;
  UpdatedBy: number;
  UpdatedDate: string;
}